Move applyOn to Standard
