+++ app/src/debug/kotlin/io/sweers/catchup/ui/debug/DebugView.kt
@@ -48,7 +48,7 @@ import io.sweers.catchup.P
 import io.sweers.catchup.R
 import io.sweers.catchup.data.LumberYard
 import io.sweers.catchup.ui.logs.LogsDialog
-import io.sweers.catchup.util.applyOn
+import io.sweers.catchup.util.kotlin.applyOn
 import io.sweers.catchup.util.d
 import io.sweers.catchup.util.isN
 import io.sweers.catchup.util.truncateAt
